requirements.txt
streamlit
pandas
plotly
yfinance
fpdf
numpy
feedparser
openpyxl
gspread
oauth2client
google-auth
toml
ib_insync